requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pytesseract==0.3.10